requirements.txt
Flask
opencv-python
torch
torchvision
dlib
google-generativeai
Pillow
numpy